FER-33 : Upgrade prettier to version 2.0.5 + Reformat code.
The HTML closing bracket is added on a new line for now (for HTML blocks/files). We will wait for prettier to create a new option (see: https://github.com/prettier/prettier/issues/5377) to fix it. It has not been planed yet so we will live without it until then.
The only files that really matter in this PR are `package.json`, `tslint.json` everything else is auto-formatting from Prettier.