FER-56, FER-107: Complete rework of fui-tabs & demo improvements
- I've added the CDK to the dev-dependencies. We'll start using it for future components and we'll need to use it for some of our current ones.
- I've completely re-do the fui-tabs component. It is working pretty much the same as (mat-tab-group)[https://v7.material.angular.io/components/tabs/overview] but without the ripple effect nor the animations.
- I've improved our demo component design and also fix FER-107 at the same time. Now, we're not using jsBeautify anymore. So be careful with your code format for demo-component. I've also removed the support for '#code' or 'class="code"' attirbute/class from the demo component as it was compiling the code within the DOM in order to extract only the code part that we where interested in.

NOTE for the future: We need to do the same as material/angular for our demo pages. Handling the code as text is NOT a good idea! Instead we should write the code examples as they where regular angular component (with HTML/TS/SCSS files and we need to extract the textual content of those files to display the code and dynamically compile them using a new demo component. See (material dev app)[https://github.com/angular/components/tree/master/src/dev-app] to better understand how to proceed.